% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_migrate.R
\name{add_stat_table_ref}
\alias{add_stat_table_ref}
\title{Title}
\usage{
add_stat_table_ref(.conn, .references, .table_ids, ..., .add_primary_key = T)
}
\arguments{
\item{.add_primary_key}{}
}
\description{
Title
}
