% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_parquet.R
\name{bind_parquet}
\alias{bind_parquet}
\title{Title}
\usage{
bind_parquet(
  .data,
  ...,
  .input_data = NULL,
  .references = get_config("references"),
  .config = getOption("rcbms.config")
)
}
\arguments{
\item{.config}{}
}
\description{
Title
}
