% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_age_gap.R
\name{check_age_gap}
\alias{check_age_gap}
\title{Title}
\usage{
check_age_gap(
  .data,
  .primary_member,
  .relation_to_primary_member,
  .threshold = 10,
  .ref_code = "relation_to_hh_head",
  .config = getOption("rcbms.config")
)
}
\arguments{
\item{.config}{}
}
\description{
Title
}
