% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_age_sex_relation.R
\name{check_age_sex_relation}
\alias{check_age_sex_relation}
\title{Title}
\usage{
check_age_sex_relation(
  .data,
  .base,
  .sex_of_head,
  .relation,
  .condition,
  .threshold
)
}
\arguments{
\item{.threshold}{}
}
\description{
Title
}
