% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicate_relation.R
\name{check_duplicate_code}
\alias{check_duplicate_code}
\title{Title}
\usage{
check_duplicate_code(
  .data,
  .var,
  .response_code,
  .config = getOption("rcbms.config")
)
}
\arguments{
\item{.config}{}
}
\description{
Title
}
