% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicate_code.R
\name{check_duplicate_relation}
\alias{check_duplicate_relation}
\title{Title}
\usage{
check_duplicate_relation(
  .data,
  .response_code,
  .config = getOption("rcbms.config")
)
}
\arguments{
\item{.config}{}
}
\description{
Title
}
