% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ethnicity.R
\name{check_ethnicity}
\alias{check_ethnicity}
\title{Title}
\usage{
check_ethnicity(
  .data,
  .var,
  .relation_var,
  .primary_member,
  .relation_to_primary_member,
  .config = getOption("rcbms.config")
)
}
\arguments{
\item{.config}{}
}
\description{
Title
}
