% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_food_insecurity.R
\name{compute_food_insecurity}
\alias{compute_food_insecurity}
\title{Title}
\usage{
compute_food_insecurity(
  .data,
  ...,
  .thres = NULL,
  .type = 1,
  .agg_levels = NULL,
  .valueset = get_config("references")$valueset,
  .extract_name_position = 5,
  .include_overall = FALSE,
  .config = getOption("rcbms.config")
)
}
\arguments{
\item{.extract_name_position}{}
}
\description{
Title
}
