% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_food_insecurity_prevalence.R
\name{compute_food_insecurity_prevalence}
\alias{compute_food_insecurity_prevalence}
\title{Title}
\usage{
compute_food_insecurity_prevalence(
  .data,
  ...,
  .results = 1,
  .agg_levels = NULL
)
}
\arguments{
\item{.agg_levels}{}
}
\description{
Title
}
