% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pgr.R
\name{compute_pgr}
\alias{compute_pgr}
\title{Title}
\usage{
compute_pgr(
  value_current,
  value_previous,
  ref_current = "2024-07-01",
  ref_previous = "2020-05-01"
)
}
\arguments{
\item{ref_previous}{}
}
\description{
Title
}
