% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_to_factor}
\alias{convert_to_factor}
\title{Title}
\usage{
convert_to_factor(
  .x,
  ...,
  .code_ref = NULL,
  .is_char = FALSE,
  .ordered = FALSE
)
}
\arguments{
\item{.ordered}{}
}
\description{
Title
}
