% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_uid.R
\name{create_barangay_geo}
\alias{create_barangay_geo}
\title{Create barangay geo ID}
\usage{
create_barangay_geo(.data, .add_length = 1)
}
\arguments{
\item{.config}{}
}
\description{
Create barangay geo ID
}
