% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_uid.R
\name{create_case_id}
\alias{create_case_id}
\title{Create case ID}
\usage{
create_case_id(
  .data,
  .filter_completed = FALSE,
  .config = getOption("rcbms.config"),
  .input_data = NULL
)
}
\arguments{
\item{.input_data}{}
}
\description{
Create case ID
}
