% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_uid.R
\name{create_line_number_id}
\alias{create_line_number_id}
\title{Create line number ID}
\usage{
create_line_number_id(.data, .join_with = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Create line number ID
}
