% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_migrate.R
\name{db_migrate}
\alias{db_migrate}
\title{Title}
\usage{
db_migrate(
  .output,
  ...,
  .config = getOption("rcbms.config"),
  .name = NULL,
  .prefix = "",
  .suffix = "",
  .add_primary_key = TRUE
)
}
\arguments{
\item{.add_primary_key}{}
}
\description{
Title
}
