% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_migrate.R
\name{db_migrate_refs}
\alias{db_migrate_refs}
\title{Title}
\usage{
db_migrate_refs(
  refs = c("data_dictionary", "macrodata", "score_card", ""),
  ...,
  .table_names = NULL,
  .include_area_names = T
)
}
\arguments{
\item{refs}{}
}
\description{
Title
}
