% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_script.R
\name{execute_script}
\alias{execute_script}
\title{Execute script}
\usage{
execute_script(survey_round, input_data, ..., config_file = NULL)
}
\arguments{
\item{...}{Additional arguments pass to \code{\link{load_required_packages}}}

\item{.config_file}{Path to config file}

\item{.survey_round}{}
}
\description{
Execute script
}
