% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_output.R
\name{generate_output}
\alias{generate_output}
\title{Generate output}
\usage{
generate_output(
  .result,
  .type = getOption("rcbms_config")$mode$type,
  .config = getOption("rcbms_config"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Generate output
}
