% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tab.R
\name{generate_tab}
\alias{generate_tab}
\title{Title}
\usage{
generate_tab(
  .data,
  .x,
  .y = NULL,
  ...,
  .agg_levels = NULL,
  .preserve_order = F,
  .top_n = NULL,
  .config = getOption("rcbms.config")
)
}
\arguments{
\item{.config}{}
}
\description{
Title
}
