% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tab_multiple.R
\name{generate_tab_multiple}
\alias{generate_tab_multiple}
\title{Title}
\usage{
generate_tab_multiple(.data, ..., .ref_vs = NULL, .agg_levels = NULL)
}
\arguments{
\item{.agg_levels}{}
}
\description{
Title
}
