% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tabs.R
\name{generate_tabs}
\alias{generate_tabs}
\title{Title}
\usage{
generate_tabs(
  .data,
  ...,
  .agg_levels = NULL,
  .total_by_cols = FALSE,
  .invert = FALSE,
  .complete = TRUE,
  .keep_col_names = FALSE,
  .multiple_response = FALSE,
  .valueset = NULL,
  .extract_name_position = 5,
  .sort = TRUE,
  .include_overall = FALSE,
  .config = getOption("rcbms.config")
)
}
\arguments{
\item{.config}{}
}
\description{
Title
}
