% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tabulation.R
\name{generate_tabulation}
\alias{generate_tabulation}
\title{Title}
\usage{
generate_tabulation(
  .ts,
  .ts_ref,
  .config,
  .save_as_excel = FALSE,
  .save_as_json = TRUE
)
}
\arguments{
\item{.save_as_json}{}
}
\description{
Title
}
