% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_validation_output.R
\name{generate_validation_output}
\alias{generate_validation_output}
\title{Generate validation output}
\usage{
generate_validation_output(
  .result,
  ...,
  .config = getOption("rcbms.config"),
  .detailed_output = F,
  .include_household_info = F,
  .add_uuid = F,
  .save_as_excel = F,
  .save_as_json = T
)
}
\arguments{
\item{.save_as_json}{}
}
\description{
Generate validation output
}
