% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_household_summary.R
\name{get_household_summary}
\alias{get_household_summary}
\title{Title}
\usage{
get_household_summary(
  .data,
  .relation_to_hh_head,
  .nuclear_family,
  .agg_levels = NULL
)
}
\arguments{
\item{.agg_levels}{}
}
\description{
Title
}
