% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_script_files.R
\name{get_script_files}
\alias{get_script_files}
\title{Get script files}
\usage{
get_script_files(.input_data, .config = getOption("rcbms.config"))
}
\arguments{
\item{.config}{}
}
\description{
Get script files
}
