% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_variable.R
\name{harmonize_variable}
\alias{harmonize_variable}
\title{Harmonize variable}
\usage{
harmonize_variable(.data, .input_data, .survey_round, .dictionary, .config)
}
\arguments{
\item{.config}{}
}
\description{
Harmonize variable
}
