% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rcbms_logs.R
\name{import_rcbms_logs}
\alias{import_rcbms_logs}
\title{Title}
\usage{
import_rcbms_logs(.dir, .user_id, .dir_to = "db-temp", .delete_source = T)
}
\arguments{
\item{.delete_source}{}
}
\description{
Title
}
