% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_and_filter_area.R
\name{join_and_filter_area}
\alias{join_and_filter_area}
\title{Title}
\usage{
join_and_filter_area(
  .data,
  ...,
  .aggregation = get_config("aggregation"),
  .config = getOption("rcbms.config"),
  .input_data = "hp",
  .retain_agg_cols = FALSE,
  .minimized = TRUE
)
}
\arguments{
\item{.minimized}{}
}
\description{
Title
}
