% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_files.R
\name{list_data_files}
\alias{list_data_files}
\title{Title}
\usage{
list_data_files(
  .input_data,
  .references = get_config("references"),
  .config = getOption("rcbms.config")
)
}
\arguments{
\item{.config}{}
}
\description{
Title
}
