% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_area_name_ref.R
\name{load_area_name_new_refs}
\alias{load_area_name_new_refs}
\title{Load new area name reference}
\usage{
load_area_name_new_refs(.gid)
}
\arguments{
\item{.gid}{}
}
\description{
Load new area name reference
}
