% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_refs.R
\name{load_data_dictionary}
\alias{load_data_dictionary}
\title{Load data dictionary}
\usage{
load_data_dictionary(.gid = get_env("DATA_DICTIONARY"), .cbms_round = NULL)
}
\arguments{
\item{.cbms_round}{}
}
\description{
Load data dictionary
}
