% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bp_data.R
\name{read_bp_data}
\alias{read_bp_data}
\title{Title}
\usage{
read_bp_data(
  .base_path,
  .references = get_config("references"),
  .config = getOption("rcbms.config")
)
}
\arguments{
\item{.config}{}
}
\description{
Title
}
