% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cbms_data.R
\name{read_cbms_data}
\alias{read_cbms_data}
\title{Rear CBMS Data}
\usage{
read_cbms_data(
  .path,
  .input_data = "hp",
  .dictionary = NULL,
  .valueset = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Rear CBMS Data
}
