% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rcdf.R
\name{read_encryped_parquet}
\alias{read_encryped_parquet}
\title{Title}
\usage{
read_encryped_parquet(
  path,
  aes_key = NULL,
  aes_iv = NULL,
  as_arrow_table = TRUE
)
}
\arguments{
\item{aes_iv}{}
}
\description{
Title
}
