% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rcdf.R
\name{read_encryped_parquet_list}
\alias{read_encryped_parquet_list}
\title{Title}
\usage{
read_encryped_parquet_list(
  path,
  input_data = NULL,
  aes_key = NULL,
  aes_iv = NULL
)
}
\arguments{
\item{aes_iv}{}
}
\description{
Title
}
