% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parquet_file.R
\name{read_parquet_file}
\alias{read_parquet_file}
\title{Title}
\usage{
read_parquet_file(
  .record,
  .input_data = "hp",
  .config = getOption("rcbms.config")
)
}
\arguments{
\item{.config}{}
}
\description{
Title
}
