% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cbms_data.R
\name{save_cbms_data}
\alias{save_cbms_data}
\title{Title}
\usage{
save_cbms_data(
  .conn,
  .df_src_files,
  .input_data,
  .pq_path,
  .p_name,
  .references,
  .config,
  .is_first_record = FALSE,
  .summary_record = NULL
)
}
\arguments{
\item{.summary_record}{}
}
\description{
Title
}
