% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_and_sort_columns.R
\name{select_and_sort_columns}
\alias{select_and_sort_columns}
\title{Title}
\usage{
select_and_sort_columns(.data, .refs, .config = getOption("rcbms_config"))
}
\arguments{
\item{.config}{}
}
\description{
Title
}
