% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_aggregation.R
\name{set_aggregation}
\alias{set_aggregation}
\title{Title}
\usage{
set_aggregation(
  ...,
  .parquet = get_config("parquet"),
  .references = get_config("references"),
  .config = getOption("rcbms.config"),
  .update_config = TRUE,
  .config_key = "aggregation"
)
}
\arguments{
\item{.config}{}
}
\description{
Title
}
