% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dot_env.R
\name{set_dot_env}
\alias{set_dot_env}
\title{Read Environment Variable}
\usage{
set_dot_env(.path = getOption("rcbms_dotenv_path"))
}
\arguments{
\item{.path}{}
}
\value{
List of environment variables defined
}
\description{
Read Environment Variable
}
\examples{
read_dot_env()

}
