% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_relative_wd}
\alias{set_relative_wd}
\title{Set current working directory relative to the project path}
\usage{
set_relative_wd(..., .base_wd = getOption("rcbms.config")$working_directory)
}
\arguments{
\item{...}{}
}
\description{
Set current working directory relative to the project path
}
