% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_logs_db.R
\name{sync_current_log}
\alias{sync_current_log}
\title{Title}
\usage{
sync_current_log(db_src, db_des, input_data, current_log_id, remarks_only = F)
}
\arguments{
\item{current_log_id}{}
}
\description{
Title
}
