% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_logs_from_src_to_des.R
\name{sync_logs_from_src_to_des}
\alias{sync_logs_from_src_to_des}
\title{Title}
\usage{
sync_logs_from_src_to_des(input_data, uid, db_src, db_des)
}
\arguments{
\item{db_des}{}
}
\description{
Title
}
