% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_rcbms_logs.R
\name{sync_rcbms_logs}
\alias{sync_rcbms_logs}
\title{Title}
\usage{
sync_rcbms_logs(
  .dir_to,
  .user_id,
  .dir_from = "//localhost/webdav/RCBMS Logs",
  .dir_temp = "rclf-temp",
  .delete_source = T
)
}
\arguments{
\item{.delete_source}{}
}
\description{
Title
}
