% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cbms_data.R
\name{tidy_cbms_data}
\alias{tidy_cbms_data}
\title{Title}
\usage{
tidy_cbms_data(
  .parquet,
  .record,
  .input_data,
  ...,
  .config = getOption("rcbms.config"),
  .complete_cases = NULL
)
}
\arguments{
\item{.complete_cases}{}
}
\description{
Title
}
