% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_geojson.R
\name{tidy_geojson}
\alias{tidy_geojson}
\title{Title}
\usage{
tidy_geojson(
  .path,
  .subfolders = c("region", "province", "city-mun", "barangay")
)
}
\arguments{
\item{.subfolders}{}
}
\description{
Title
}
