% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_area_name.R
\name{transform_area_name}
\alias{transform_area_name}
\title{Transform area name}
\usage{
transform_area_name(.area_name, .add_length = 0)
}
\arguments{
\item{.add_length}{}
}
\description{
Transform area name
}
