% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Title}
\usage{
validate(
  .data,
  .id,
  ...,
  .refs = NULL,
  .title = get_from_validation_id(.id, .refs, "title"),
  .description = get_from_validation_id(.id, .refs, "description"),
  .section = get_from_validation_id(.id, .refs, "section"),
  .priority_level = get_from_validation_id(.id, .refs, "priority_level"),
  .primary_data_item = get_from_validation_id(.id, .refs, "primary_data_item")
)
}
\arguments{
\item{.primary_data_item}{}
}
\description{
Title
}
