% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_age_sex_relation.R
\name{check_age_sex_relation}
\alias{check_age_sex_relation}
\title{Title}
\usage{
check_age_sex_relation(
  .data,
  .primary_member,
  .sex_of_primary_member = c(1, 2),
  .relation_to_primary_member,
  .threshold,
  .condition = "<",
  .config = getOption("rcbms.config")
)
}
\arguments{
\item{.config}{}
}
\description{
Title
}
