% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_age_sex_relation.R
\name{check_age_sex_relation}
\alias{check_age_sex_relation}
\title{Check sex and age based on relationship to household head}
\usage{
check_age_sex_relation(
  .data,
  .primary_member,
  .sex_of_primary_member = c(1, 2),
  .relation_to_primary_member,
  .threshold,
  .condition = "<",
  .conjuction = NULL,
  .relation_to_hh_head_var = NULL,
  .config = getOption("rcbms.config")
)
}
\arguments{
\item{.config}{}
}
\description{
Check sex and age based on relationship to household head
}
