% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_to_factor}
\alias{convert_to_factor}
\title{Title}
\usage{
convert_to_factor(
  ...,
  code_ref = NULL,
  input_data = "hp",
  is_char = F,
  ordered = F
)
}
\arguments{
\item{ordered}{}
}
\description{
Title
}
