% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_migrate.R
\name{db_migrate}
\alias{db_migrate}
\title{Title}
\usage{
db_migrate(
  .output,
  ...,
  .name = NULL,
  .prefix = "",
  .suffix = "",
  .add_primary_key = TRUE,
  .references = get_config("references")$macrodata,
  .config = getOption("rcbms.config")
)
}
\arguments{
\item{.config}{}
}
\description{
Title
}
