% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_script.R
\name{execute_script}
\alias{execute_script}
\title{Execute script}
\usage{
execute_script(.config_file, ...)
}
\arguments{
\item{.config_file}{Path to config file}

\item{...}{Additional arguments passed to \code{\link{load_required_packages}}}
}
\description{
Execute script
}
