% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_script.R
\name{execute_script}
\alias{execute_script}
\title{Title}
\usage{
execute_script(
  .parquet,
  .references,
  .aggregation,
  .config = getOption("rcbms_config"),
  .excluded_cases = NULL
)
}
\arguments{
\item{.config}{}
}
\description{
Title
}
