% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_script.R
\name{execute_script}
\alias{execute_script}
\title{Title}
\usage{
execute_script(
  .parquet,
  .refs,
  .aggregation,
  .config = getOption("rcbms_config"),
  .filter_complete_cases = T
)
}
\arguments{
\item{.filter_complete_cases}{}
}
\description{
Title
}
