% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_validation.R
\name{generate_validation}
\alias{generate_validation}
\title{Generate validation output}
\usage{
generate_validation(
  .cv,
  .references,
  .config,
  .detailed_output = FALSE,
  .include_additional_info = FALSE,
  .add_uuid = FALSE,
  .save_as_excel = FALSE,
  .save_as_json = TRUE
)
}
\arguments{
\item{.save_as_json}{}
}
\description{
Generate validation output
}
