% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_validation.R
\name{generate_validation}
\alias{generate_validation}
\title{Generate validation output}
\usage{
generate_validation(.cv, .cv_ref, .config)
}
\arguments{
\item{.config}{}
}
\description{
Generate validation output
}
