% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_validation_par.R
\name{generate_validation_par}
\alias{generate_validation_par}
\title{Generate validation output in parallel mode}
\usage{
generate_validation_par(.cv, .cv_ref, .config)
}
\arguments{
\item{.config}{}
}
\description{
Generate validation output in parallel mode
}
