% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_complete_cases.R
\name{get_complete_cases}
\alias{get_complete_cases}
\title{Title}
\usage{
get_complete_cases(
  .parquet,
  ...,
  .config = getOption("rcbms_config"),
  .excluded_cases = NULL,
  .filter_by_area = FALSE
)
}
\arguments{
\item{.excluded_cases}{}
}
\description{
Title
}
