% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_household_summary.R
\name{get_household_summary}
\alias{get_household_summary}
\title{Title}
\usage{
get_household_summary(
  .data,
  .sex,
  .relation_to_hh_head,
  .nuclear_family,
  .agg_levels = NULL,
  .include_overall = T,
  .include_survey_round = T,
  .include_individuals = T,
  .config = getOption("rcbms.config")
)
}
\arguments{
\item{.config}{}
}
\description{
Title
}
