% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_script_files.R
\name{get_script_files}
\alias{get_script_files}
\title{Get script files}
\usage{
get_script_files(.config = getOption("rcbms_config"), .input_data = "hp")
}
\arguments{
\item{.input_data}{}
}
\description{
Get script files
}
