% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_variable.R
\name{harmonize_variable}
\alias{harmonize_variable}
\title{Harmonize variable}
\usage{
harmonize_variable(
  .data,
  .dictionary,
  .survey_round,
  .input_data,
  .config = getOption("rcbms.config")
)
}
\arguments{
\item{.config}{}
}
\description{
Harmonize variable
}
