% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_references.R
\name{load_references}
\alias{load_references}
\title{Title}
\usage{
load_references(
  .config = getOption("rcbms.config"),
  .update_config = TRUE,
  .config_key = "references"
)
}
\arguments{
\item{.config_key}{}
}
\description{
Title
}
