% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cbms_data.R
\name{read_cbms_data}
\alias{read_cbms_data}
\title{Title}
\usage{
read_cbms_data(
  .references = get_config("references"),
  .config = getOption("rcbms.config"),
  .assign_name = "parquet"
)
}
\arguments{
\item{.assign_name}{}
}
\description{
Title
}
