% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_signature_file.R
\name{read_signature_files}
\alias{read_signature_files}
\title{Title}
\usage{
read_signature_files(
  .dir,
  .area_code = NULL,
  .sig_folder = NULL,
  .formats = c("png", "jpg", "jpeg"),
  .options = list(ink_threshold = 375, img_threshold = 1e+05, crop_threshold = 500,
    filename_length = 31)
)
}
\arguments{
\item{.options}{}
}
\description{
Title
}
