% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_aggregation.R
\name{set_aggregation}
\alias{set_aggregation}
\title{Title}
\usage{
set_aggregation(
  .parquet,
  .refs,
  ...,
  .config = getOption("rcbms_config"),
  .input_data = "hp"
)
}
\arguments{
\item{.config}{}
}
\description{
Title
}
