% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_config.R
\name{set_config}
\alias{set_config}
\title{Set initial configuration for the project}
\usage{
set_config(
  .config_dir = "configs",
  .survey_round = NULL,
  .input_data = NULL,
  .include_env = TRUE
)
}
\arguments{
\item{.config_dir}{Path to directory containing valid configuration files. Default config files will be generated if the directory specified does not exist.}

\item{.survey_round}{Survey round which points to the \code{src} folder of your project directory.}

\item{.input_data}{Input data which, by default, points to the \code{src} > \code{data} folder of your project directory.}

\item{.include_env}{Include \code{.env} file. Default is \code{TRUE}. This will be switched to \code{FALSE} if the \code{.env} file is not present. The default location is \code{"./configs/.env"}.}
}
\value{
A list of configuration settings use to execute the script.
}
\description{
Set initial configuration for the project
}
\examples{

set_config("./configs/global.yml")

}
