% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_config.R
\name{set_config}
\alias{set_config}
\title{Title}
\usage{
set_config(
  .config_file,
  .include_env = TRUE,
  .save_as_options = TRUE,
  .assign_name = "config"
)
}
\arguments{
\item{.config_file}{Path to config file}

\item{.assign_name}{}
}
\description{
Title
}
