% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_config.R
\name{set_config}
\alias{set_config}
\title{Set initial configuration for the project}
\usage{
set_config(
  .config_dir = "configs",
  .survey_round = NULL,
  .input_data = NULL,
  .include_env = TRUE
)
}
\arguments{
\item{.include_env}{Include \code{.env} file. Default is \code{TRUE}. This will be switched to \code{FALSE} if the \code{.env} file is not present. The default location is \code{"./configs/.env"}.}

\item{.config_file}{Path to config file. Default is \code{"./configs/global.yml"}. If the file is not present, a default config file will be generated.}
}
\value{
A list of configuration settings use to execute the script.
}
\description{
Set initial configuration for the project
}
\examples{

set_config("./configs/global.yml")

}
