% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Title}
\usage{
validate(
  .data,
  .id,
  ...,
  .references = get_config("references"),
  .title = get_from_validation_id(.id, .references, "title"),
  .description = get_from_validation_id(.id, .references, "description"),
  .section = get_from_validation_id(.id, .references, "section"),
  .priority_level = get_from_validation_id(.id, .references, "priority_level"),
  .primary_data_item = get_from_validation_id(.id, .references, "primary_data_item"),
  .validation_type = get_from_validation_id(.id, .references, "validation_type"),
  .status = get_from_validation_id(.id, .references, "status"),
  .date_introduced = get_from_validation_id(.id, .references, "date_introduced")
)
}
\arguments{
\item{.date_introduced}{}
}
\description{
Title
}
